requirements.txt
fastapi
sqlmodel
guvicorn